// src/themes/types.ts
export interface Theme {
    colors: {
        anotherPurple: string
        purple: string
        littlePurple: string
        purpleDark: string
        green: string
        anotherGreen: string
        delete: string
    }
    back: {
        linesInWhite: string
        hover: string
        background: string
    }
    shapes: {
        '01': string
        '02': string
        disabled: string
    }
    texts: {
        titles: string
        base: string
        complements: string
        inputs: string
    }
    textsInPurple: {
        title: string
        base: string
        complement: string
    }
}